refactoring and making tests only run when a token is present
